refactor: remove ChooseProtocolView and fix network setup router
