+++ packages/desktop/views/onboarding/views/network-setup/views/index.js
@@ -1,3 +1,2 @@
 export { default as ChooseNetworkView } from './ChooseNetworkView.svelte'
-export { default as ChooseProtocolView } from './ChooseProtocolView.svelte'
-export { default as SetupPrivateNetworkConnectionView } from './SetupPrivateNetworkConnectionView.svelte'
+export { default as SetupCustomNetworkConnectionView } from './SetupCustomNetworkConnectionView.svelte'
