+++ packages/shared/lib/core/router/enums/onboarding/network-setup-route.enum.ts
@@ -1,5 +1,4 @@
 export enum NetworkSetupRoute {
     ChooseNetwork = 'chooseNetwork',
-    ChooseProtocol = 'chooseProtocol',
-    SetupPrivateNetworkConnection = 'setupPrivateNetworkConnection',
+    SetupCustomNetworkConnection = 'setupCustomNetworkConnection',
 }
