+++ packages/shared/lib/core/router/subrouters/onboarding/network-setup-router.ts
@@ -1,8 +1,6 @@
-import { get, writable } from 'svelte/store'
-
 import { onboardingProfile } from '@contexts/onboarding'
-import { NetworkType } from '@core/network'
-
+import { NetworkId } from '@core/network'
+import { get, writable } from 'svelte/store'
 import { Subrouter } from '../../classes'
 import { NetworkSetupRoute } from '../../enums'
 import { onboardingRouter } from '../../routers'
@@ -21,27 +19,17 @@ export class NetworkSetupRouter extends Subrouter<NetworkSetupRoute> {
         const _onboardingProfile = get(onboardingProfile)
         const currentRoute = get(this.routeStore)
         switch (currentRoute) {
-            case NetworkSetupRoute.ChooseProtocol: {
-                const isDeveloperProfile = _onboardingProfile?.isDeveloperProfile
-                if (isDeveloperProfile) {
-                    nextRoute = NetworkSetupRoute.ChooseNetwork
-                    break
-                } else {
-                    this.parentRouter.next()
-                    return
-                }
-            }
             case NetworkSetupRoute.ChooseNetwork: {
-                const networkType = _onboardingProfile?.networkType ?? NetworkType.Devnet
-                if (networkType === NetworkType.PrivateNet) {
-                    nextRoute = NetworkSetupRoute.SetupPrivateNetworkConnection
+                const networkId = _onboardingProfile?.network ?? NetworkId.Custom
+                if (networkId === NetworkId.Custom) {
+                    nextRoute = NetworkSetupRoute.SetupCustomNetworkConnection
                     break
                 } else {
                     this.parentRouter.next()
                     return
                 }
             }
-            case NetworkSetupRoute.SetupPrivateNetworkConnection:
+            case NetworkSetupRoute.SetupCustomNetworkConnection:
                 this.parentRouter.next()
                 return
         }
